user model & auth

+++ medsos_API/controller/feed.js
@@ -4,6 +4,7 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
     const perPage = 2;
@@ -64,18 +65,23 @@ exports.createPost = (req, res, next) => {
     const post = new Post({
         title: title,
         content: content,
-        creator: {
-            name: 'Iqbal'
-        },
-        imageUrl: imageUrl
+        imageUrl: imageUrl,
+        creator: req.userId
     });
-
+    let creator;
     post.save()
         .then(result => {
-            console.log(result);
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        })
+        .then(result => {
             res.status(201).json({
                 message: 'Post created successfully',
-                post: result
+                post: post
             })
         })
         .catch(err => {
@@ -114,6 +120,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw (error);
             }
+            if (req.userId.toString() !== post.creator.toString()) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw (error);
+            }
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl)
             }
@@ -143,12 +154,22 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw (error);
             }
-            //check logged in user
+            if (req.userId.toString() !== post.creator.toString()) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw (error);
+            }
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
         .then(result => {
-            console.log('deleted', result);
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then(result => {
             return res.status(200).json({ message: 'Post is deleted.' });
         })
         .catch(err => {
